use bfs for finding entities of a certain type

+++ source/simulation/jobs/path_finder.h
@@ -11,10 +11,9 @@ class PathFinder : public Job {
       PathFinder(Simulation& sim, std::string name) : Job(sim, name) { }
       virtual ~PathFinder() { };
 
-      virtual int EstimateCostToSolution() = 0;
+      virtual float EstimateCostToSolution() = 0;
 };
 
 END_RADIANT_SIMULATION_NAMESPACE
 
 #endif // _RADIANT_SIMULATION_JOBS_PATH_FINDER_H
-
